Refactor catalog to self document
