+++ internal/sql/catalog/schema.go
@@ -2,10 +2,93 @@ package catalog
 
 import (
 	"fmt"
+	"strings"
 
 	"github.com/kyleconroy/sqlc/internal/sql/ast"
 	"github.com/kyleconroy/sqlc/internal/sql/sqlerr"
 )
+
+type Schema struct {
+	Name   string
+	Tables []*Table
+	Types  []Type
+	Funcs  []*Function
+
+	Comment string
+}
+
+func (s *Schema) getFunc(rel *ast.FuncName, tns []*ast.TypeName) (*Function, int, error) {
+	for i := range s.Funcs {
+		if strings.ToLower(s.Funcs[i].Name) != strings.ToLower(rel.Name) {
+			continue
+		}
+
+		args := s.Funcs[i].InArgs()
+		if len(args) != len(tns) {
+			continue
+		}
+		found := true
+		for j := range args {
+			if !sameType(s.Funcs[i].Args[j].Type, tns[j]) {
+				found = false
+				break
+			}
+		}
+		if !found {
+			continue
+		}
+		return s.Funcs[i], i, nil
+	}
+	return nil, -1, sqlerr.RelationNotFound(rel.Name)
+}
+
+func (s *Schema) getFuncByName(rel *ast.FuncName) (*Function, int, error) {
+	idx := -1
+	name := strings.ToLower(rel.Name)
+	for i := range s.Funcs {
+		lowered := strings.ToLower(s.Funcs[i].Name)
+		if lowered == name && idx >= 0 {
+			return nil, -1, sqlerr.FunctionNotUnique(rel.Name)
+		}
+		if lowered == name {
+			idx = i
+		}
+	}
+	if idx < 0 {
+		return nil, -1, sqlerr.RelationNotFound(rel.Name)
+	}
+	return s.Funcs[idx], idx, nil
+}
+
+func (s *Schema) getTable(rel *ast.TableName) (*Table, int, error) {
+	for i := range s.Tables {
+		if s.Tables[i].Rel.Name == rel.Name {
+			return s.Tables[i], i, nil
+		}
+	}
+	return nil, -1, sqlerr.RelationNotFound(rel.Name)
+}
+
+func (s *Schema) getType(rel *ast.TypeName) (Type, int, error) {
+	for i := range s.Types {
+		switch typ := s.Types[i].(type) {
+		case *Enum:
+			if typ.Name == rel.Name {
+				return s.Types[i], i, nil
+			}
+		}
+	}
+	return nil, -1, sqlerr.TypeNotFound(rel.Name)
+}
+
+func (c *Catalog) getSchema(name string) (*Schema, error) {
+	for i := range c.Schemas {
+		if c.Schemas[i].Name == name {
+			return c.Schemas[i], nil
+		}
+	}
+	return nil, sqlerr.SchemaNotFound(name)
+}
 
 func (c *Catalog) createSchema(stmt *ast.CreateSchemaStmt) error {
 	if stmt.Name == nil {
